+++ src/main/java/com/gsm/blabla/auth/application/AuthService.java
@@ -6,14 +6,13 @@ import com.gsm.blabla.global.response.Code;
 import com.gsm.blabla.global.exception.GeneralException;
 import com.gsm.blabla.jwt.TokenProvider;
 import com.gsm.blabla.jwt.application.JwtService;
-import com.gsm.blabla.jwt.dao.AppleAccountRepository;
-import com.gsm.blabla.jwt.dao.GoogleAccountRepository;
+import com.gsm.blabla.auth.dao.AppleAccountRepository;
+import com.gsm.blabla.auth.dao.GoogleAccountRepository;
 import com.gsm.blabla.jwt.dao.JwtRepository;
-import com.gsm.blabla.jwt.domain.AppleAccount;
-import com.gsm.blabla.jwt.domain.GoogleAccount;
+import com.gsm.blabla.auth.domain.AppleAccount;
+import com.gsm.blabla.auth.domain.GoogleAccount;
 import com.gsm.blabla.jwt.domain.Jwt;
 import com.gsm.blabla.jwt.dto.AppleAccountDto;
-import com.gsm.blabla.jwt.dto.ApplePublicKeyDto;
 import com.gsm.blabla.jwt.dto.AppleTokenDto;
 import com.gsm.blabla.jwt.dto.GoogleAccountDto;
 import com.gsm.blabla.jwt.dto.JwtDto;
