+++ src/main/java/com/gsm/blabla/auth/dao/AppleAccountRepository.java
@@ -1,6 +1,6 @@
-package com.gsm.blabla.jwt.dao;
+package com.gsm.blabla.auth.dao;
 
-import com.gsm.blabla.jwt.domain.AppleAccount;
+import com.gsm.blabla.auth.domain.AppleAccount;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
