+++ src/main/java/com/gsm/blabla/auth/dao/GoogleAccountRepository.java
@@ -1,6 +1,6 @@
-package com.gsm.blabla.jwt.dao;
+package com.gsm.blabla.auth.dao;
 
-import com.gsm.blabla.jwt.domain.GoogleAccount;
+import com.gsm.blabla.auth.domain.GoogleAccount;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
